requirements.txt
yfinance==0.2.51
pandas==2.2.3
gspread==6.1.4
google-auth==2.37.0
pdfplumber==0.11.4
requests==2.32.3
python-dateutil==2.9.0
